requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.7
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==2.1.8
django-allauth==0.41.0
django-celery==3.3.1
django-celery-beat==2.0.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
importlib-metadata==1.6.0
kombu==3.0.37
Markdown==3.2.2
oauthlib==3.1.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.4.2
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
redis==3.5.2
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
